--- a/JetMETCorrections/Type1MET/plugins/PFCandMETcorrInputProducer.cc
+++ b/JetMETCorrections/Type1MET/plugins/PFCandMETcorrInputProducer.cc
@@ -12,29 +12,25 @@ PFCandMETcorrInputProducer::PFCandMETcorrInputProducer(const edm::ParameterSet& 
     vParameterSet cfgBinning = cfg.getParameter<vParameterSet>("binning");
     for ( vParameterSet::const_iterator cfgBinningEntry = cfgBinning.begin();
 	  cfgBinningEntry != cfgBinning.end(); ++cfgBinningEntry ) {
-      binning_.push_back(new binningEntryType(*cfgBinningEntry));
+      binning_.emplace_back(new binningEntryType(*cfgBinningEntry));
     }
   } else {
-    binning_.push_back(new binningEntryType());
+    binning_.emplace_back(new binningEntryType());
   }
   
-  for ( std::vector<binningEntryType*>::const_iterator binningEntry = binning_.begin();
+  for ( auto binningEntry = binning_.begin();
 	binningEntry != binning_.end(); ++binningEntry ) {
     produces<CorrMETData>((*binningEntry)->binLabel_);
   }
 }
 
 PFCandMETcorrInputProducer::~PFCandMETcorrInputProducer()
-{
-  for ( std::vector<binningEntryType*>::const_iterator it = binning_.begin();
-	it != binning_.end(); ++it ) {
-    delete (*it);
-  }
+{  
 }
 
 void PFCandMETcorrInputProducer::produce(edm::Event& evt, const edm::EventSetup& es)
 {
-  for ( std::vector<binningEntryType*>::iterator binningEntry = binning_.begin();
+  for ( auto binningEntry = binning_.begin();
 	binningEntry != binning_.end(); ++binningEntry ) {
     (*binningEntry)->binUnclEnergySum_ = CorrMETData();
   }
@@ -45,7 +41,7 @@ void PFCandMETcorrInputProducer::produce(edm::Event& evt, const edm::EventSetup&
   
   for ( edm::View<reco::Candidate>::const_iterator cand = cands->begin();
 	cand != cands->end(); ++cand ) {
-    for ( std::vector<binningEntryType*>::iterator binningEntry = binning_.begin();
+    for ( auto binningEntry = binning_.begin();
 	  binningEntry != binning_.end(); ++binningEntry ) {
       if ( !(*binningEntry)->binSelection_ || (*(*binningEntry)->binSelection_)(cand->p4()) ) {
 	(*binningEntry)->binUnclEnergySum_.mex   += cand->px();
@@ -56,8 +52,8 @@ void PFCandMETcorrInputProducer::produce(edm::Event& evt, const edm::EventSetup&
   }
 
 //--- add momentum sum of PFCandidates not within jets ("unclustered energy") to the event
-  for ( std::vector<binningEntryType*>::const_iterator binningEntry = binning_.begin();
-	binningEntry != binning_.end(); ++binningEntry ) {
+  for ( auto binningEntry = binning_.cbegin();
+	binningEntry != binning_.cend(); ++binningEntry ) {
     evt.put(std::auto_ptr<CorrMETData>(new CorrMETData((*binningEntry)->binUnclEnergySum_)), (*binningEntry)->binLabel_);
   }
 }