… PhysicsTools/JetMCAlgos/plugins/JetFlavourClustering.cc
… PhysicsTools/JetMCAlgos/plugins/TauGenJetProducer.cc
… PhysicsTools/JetMCAlgos/plugins/TauGenJetProducer.h
… PhysicsTools/PatAlgos/interface/EfficiencyLoader.h
… PhysicsTools/PatAlgos/interface/KinResolutionsLoader.h
… PhysicsTools/PatAlgos/plugins/CandidateSummaryTable.cc
… PhysicsTools/PatAlgos/plugins/DuplicatedElectronCleaner.cc
… PhysicsTools/PatAlgos/plugins/JetCorrFactorsProducer.cc
… PhysicsTools/PatAlgos/plugins/JetCorrFactorsProducer.h
… PhysicsTools/PatAlgos/plugins/PATCleaner.h
… PhysicsTools/PatAlgos/plugins/PATCompositeCandidateProducer.cc
… PhysicsTools/PatAlgos/plugins/PATCompositeCandidateProducer.h
… PhysicsTools/PatAlgos/plugins/PATConversionProducer.cc
… PhysicsTools/PatAlgos/plugins/PATConversionProducer.h
… PhysicsTools/PatAlgos/plugins/PATElectronProducer.cc
… PhysicsTools/PatAlgos/plugins/PATElectronProducer.h
… PhysicsTools/PatAlgos/plugins/PATElectronSlimmer.cc
… PhysicsTools/PatAlgos/plugins/PATGenCandsFromSimTracksProducer.cc
… PhysicsTools/PatAlgos/plugins/PATGenJetSlimmer.cc
… PhysicsTools/PatAlgos/plugins/PATGenericParticleProducer.h
… PhysicsTools/PatAlgos/plugins/PATHeavyIonProducer.cc
… PhysicsTools/PatAlgos/plugins/PATHemisphereProducer.cc
… PhysicsTools/PatAlgos/plugins/PATHemisphereProducer.h
… PhysicsTools/PatAlgos/plugins/PATJetProducer.h
… PhysicsTools/PatAlgos/plugins/PATJetSelector.h
… PhysicsTools/PatAlgos/plugins/PATJetSlimmer.cc
… PhysicsTools/PatAlgos/plugins/PATJetUpdater.h
… PhysicsTools/PatAlgos/plugins/PATLeptonCountFilter.cc
… PhysicsTools/PatAlgos/plugins/PATLeptonCountFilter.h
… PhysicsTools/PatAlgos/plugins/PATLostTracks.cc