Using igprof I found and converted the last remaining legacy reco modules that were taking some amount of time.

I now see no edm::EDProducers being run when running RECO+MiniAOD only in wf25202.

This PR is based on #10620, I'd prefer to keep them separate since this PR concerns RECO rather than MiniAOD.

@Dr15Jones The last five commits in the branch are the new ones. The conversions were pretty straightforward as far as I could tell.