@cmsbuild 2015-08-16
A new Pull Request was created by @lgray (Lindsey Gray) for CMSSW_7_6_X.

Convert remaining modules in RECO to modern modules (76X)

It involves the following packages:

CommonTools/ParticleFlow
CommonTools/RecoAlgos
CommonTools/RecoUtils
CommonTools/UtilAlgos
CommonTools/Utils
DataFormats/PatCandidates
JetMETCorrections/Type1MET
L1Trigger/L1TCalorimeter
PhysicsTools/HepMCCandAlgos
PhysicsTools/IsolationAlgos
PhysicsTools/JetCharge
PhysicsTools/JetMCAlgos
PhysicsTools/PatAlgos
PhysicsTools/PatUtils
RecoEgamma/EgammaPhotonProducers
RecoJets/JetProducers
RecoLocalCalo/HcalRecProducers
RecoTauTag/RecoTau


@cvuosalo, @monttj, @cmsbuild, @slava77, @vadler, @mulhearn can you please review it and eventually sign? Thanks.
@rappoccio, @Sam-Harper, @jdolen, @imarches, @makortel, @acaudron, @abbiendi, @mmarionncern, @jhgoh, @Martin-Grunewald, @ahinzmann, @nhanvtran, @yslai, @schoef, @ferencek, @gpetruc, @mariadalfonso, @pvmulder, @TaiSakuma this is something you requested to watch as well.
You can sign-off by replying to this message having '+1' in the first line of your reply.
You can reject by replying  to this message having '-1' in the first line of your reply.
If you are a L2 or a release manager you can ask for tests by saying 'please test' in the first line of a comment.
@degano you are the release manager for this.
You can merge this pull request by typing 'merge' in the first line of your comment.