@lgray 2015-08-16
@cmsbuild please test