@cmsbuild 2015-08-16
The tests are being triggered in jenkins.