@slava77 2015-08-16
@lgray can you run helgrind on this?

```
valgrind --tool=helgrind
--suppressions=$CMSSW_RELEASE_BASE/src/Utilities/ReleaseScripts/data/cms-valgrind-helgrind.supp
--num-callers=20 --xml=yes --xml-file=helgrind-%p.xml cmsRun a.py >& a.log
```