@lgray 2015-08-16
@slava77 Sure. How many events should I run, how many threads?