@slava77 2015-08-16
(out of thin air)
8 threads 100 events
- 8 just to stress it a bit more than the "default" 4 we use in production/prompt
- 100 so that you get an answer sooner than Tuesday