@Dr15Jones 2015-08-16
I usually do 4 threads with 10 events since most problems happen on the first events. You also need to use the valgrind option which allows use with a JIT because of ROOT6. Fortunately I can't look that up right now.