@cmsbuild 2015-08-16
-1 
 
Tested at: bae4a14fd2a3ba8adb487b4e8e5ea56865a51eff
I found errors in the following unit tests: 
 
 ---> test runtestPhysicsToolsPatAlgos had ERRORS
 
 
you can see the results of the tests here: 
 https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-10806/7369/summary.html