@lgray 2015-08-16
@Dr15Jones @slava77 FYI the option for running with JITs seems to be ```--smc-check=all```