@slava77 2015-08-16
ah, yes, I knew that (the command above was a copy-paste from root5 times)