@cmsbuild 2015-08-16
Comparison is ready
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-10806/7369/summary.html