@lgray 2015-08-17
@slava77 @Dr15Jones Ran helgrind, I get an obscene number of "possible data race during read" errors, many of them appearing in the tracking code. Is there anything I should particularly be watching out for?