@lgray 2015-08-17
There are also "possible data race during write" errors, in fastjet and many other places. However, there is no contextual information provided to figure out if it is really a problem.