@lgray 2015-08-17
More info (after reading a little):
- There are 188 instances of helgrind actually emitting a ```This conflicts with a previous write of size``` 
   - Some of these are within tbb itself
   - The MagneticField constructor
   - Track trajectory state
   - HcalHitReconstructor::produce (either minuit2 or HcalCalibrations it looks like?)
   - There are more....

Most notably I don't see anything prefixed with PAT, nor do I see any of the modules in the last five commits.

The helgrind file can be found here on cern afs:
~lgray/work/public/CMSSW_7_6_X_2015-08-14-1100/src/matrix_test/25202.0_TTbar_13+TTbar_13INPUT+DIGIUP15_PU25+RECOUP15_PU25+HARVESTUP15_PU25/helgrind-26489.xml

Please take a look and let me know if you find anything.