@Dr15Jones 2015-08-17
As for helgrind, it unfortunately does not understand atomic and incorrectly marks its use as possibly thread unsafe.