@Dr15Jones 2015-08-17
The helgrind report does show a real race condition:
```
  <xwhat>
    <text>Possible data race during write of size 8 at 0x3042F910 by thread #2</text>
    <hthreadid>2</hthreadid>
  </xwhat>
  <stack>
    <frame>
      <ip>0x30F16989</ip>
      <obj>/afs/cern.ch/cms/sw/ReleaseCandidates/volA/slc6_amd64_gcc493/cms/cmssw-patch/CMSSW_7_6_X_2015-08-14-1100/lib/slc6_amd64_gcc493/pluginPhysicsToolsUtilAlgos_plugins.so</obj>
      <fn>edm::FilterObjectWrapper&lt;PVObjectSelector, std::vector&lt;reco::Vertex, std::allocator&lt;reco::Vertex&gt; &gt; &gt;::filter(edm::StreamID, edm::Event&amp;, edm::EventSetup const&amp;) const</fn>
    </frame>
...
  <auxwhat>Address 0x3042f910 is 0 bytes inside a block of size 8 alloc'd</auxwhat>
  <stack>
    <frame>
      <ip>0x40294A7</ip>
      <obj>/afs/cern.ch/cms/sw/ReleaseCandidates/volA/slc6_amd64_gcc493/external/valgrind/3.10.1/lib/valgrind/vgpreload_helgrind-amd64-linux.so</obj>
      <fn>operator new(unsigned long)</fn>
    </frame>
    <frame>
      <ip>0x30F13765</ip>
      <obj>/afs/cern.ch/cms/sw/ReleaseCandidates/volA/slc6_amd64_gcc493/cms/cmssw-patch/CMSSW_7_6_X_2015-08-14-1100/lib/slc6_amd64_gcc493/pluginPhysicsToolsUtilAlgos_plugins.so</obj>
      <fn>PVObjectSelector::PVObjectSelector(edm::ParameterSet const&amp;)</fn>
    </frame>
```

The problem is ```FilterObjectWrapper``` was made into a global module
http://cmslxr.fnal.gov/lxr/source/PhysicsTools/UtilAlgos/interface/EDFilterObjectWrapper.h?v=CMSSW_7_6_X_2015-08-16-2300#0049

but the ```filter_``` it holds is non-const
http://cmslxr.fnal.gov/lxr/source/PhysicsTools/UtilAlgos/interface/EDFilterObjectWrapper.h?v=CMSSW_7_6_X_2015-08-16-2300#0095

 and in this case changes state
http://cmslxr.fnal.gov/source/PhysicsTools/SelectorUtils/interface/PVObjectSelector.h?v=CMSSW_7_6_X_2015-08-13-2300

Therefore the ```FilterObjectWrapper``` must be switched to at very least a ```stream``` type module.