@cvuosalo 2015-08-26
+1

For #10806 1a151b79bb33e96e8450c406f1a0b56aac586889

Conversion of RECO modules to support multi-threading. There should be no changes in monitored quantities.

The code changes are satisfactory, and Jenkins [tests](https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-10806/7619/summary.html) against baseline CMSSW_7_6_X_2015-08-24-1100 show no significant differences, as expected. Testing workflow 25202.0_TTbar_13 with 70 events against baseline CMSSW_7_6_X_2015-08-24-1100, with four threads for both and no DQM or Validation, shows about a 0.5% increase in memory use and about a 3% decrease in processing time:
```
Baseline
Max VSIZ 2805.5 on evt 65 ; max RSS 2308.01 on evt 53
Time av 16.2831 s/evt   max 33.3318 s on evt 30
M1 Time av 16.056 s/evt   max 33.3318 s on evt 30

PR 10806 (plus 10805 and 10873)
Max VSIZ 2827.61 on evt 65 ; max RSS 2320.44 on evt 53
Time av 15.8305 s/evt   max 33.3892 s on evt 30
M1 Time av 15.618 s/evt   max 33.3892 s on evt 30
```
Another timing measurement, baseline ==> PR:
```
Job total:  15.9386 s/ev ==> 15.6316 s/ev
Job total:  15.4252 s/ev ==> 15.197 s/ev (with first event excluded)
```